requirements.txt
numpy>=1.22
pygame>=2.1
PyYAML>=6.0
matplotlib>=3.5
pandas>=1.4